Fix webpack bundles, reorganize css and js files
