+++ saleor/static/scss/dashboard/dashboard.scss
