+++ saleor/static/scss/dashboard/widget/_breadcrumbs.scss
