+++ saleor/static/scss/dashboard/widget/_card.scss
