+++ saleor/static/scss/dashboard/widget/_data-table.scss
