+++ saleor/static/scss/dashboard/widget/_dropzone.scss
