+++ saleor/static/scss/dashboard/widget/_form.scss
