+++ saleor/static/scss/dashboard/widget/_list.scss
