+++ saleor/static/scss/dashboard/widget/_ppoi.scss
