+++ saleor/static/scss/dashboard/widget/_product-gallery.scss
