+++ saleor/static/scss/dashboard/widget/_select2.scss
