+++ saleor/static/scss/dashboard/widget/_tabs.scss
