+++ saleor/static/scss/dashboard/widget/_toast.scss
