--- saleor/static/scss/storefront.scss
@@ -1,186 +1,0 @@
-$icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
-@import "~bootstrap-sass/assets/stylesheets/bootstrap";
-
-.table > tbody > tr > td.noborder {
-  border-top: 1px solid transparent;
-}
-
-.navbar {
-	margin-bottom: $grid-gutter-width;
-}
-
-.price {
-	&.undiscounted {
-		position: relative;
-    &:before {
-      border-top: 1px solid;
-      border-color: inherit;
-      content: "";
-      left: 0;
-      position: absolute;
-      right: 0;
-      top: 50%;
-      transform: rotate(-5deg);
-    }
-	}
-}
-
-.address-details {
-	border-bottom: 1px dashed;
-	color: #3a87ad;
-	cursor: pointer;
-}
-
-.i18n-address {
-  .row {
-    @media (min-width: $screen-sm-min) {
-      display: flex;
-
-      .col-sm-4 {
-        float: none;
-        flex: 1;
-      }
-    }
-  }
-}
-
-.popover {
-	min-width: 200px;
-}
-
-.product-list-item-name {
-	display: block;
-	margin-top: $grid-gutter-width/4;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.product-gallery {
-	margin-top: $grid-gutter-width;
-}
-
-.product-related {
-	.active {
-		border: 1px solid $btn-primary-border;
-		box-shadow: 0 0 5px $btn-primary-border;
-	}
-}
-
-.product-form {
-	.btn-group {
-		display: block;
-		overflow: hidden;
-	}
-}
-
-.breadcrumb {
-	> li+li {
-		&:before {
-			content: '/';
-		}
-	}
-}
-
-.nav {
-	> li {
-		> a {
-			@extend .text-info;
-		}
-
-		&.active {
-			> a, a:hover {
-				background-color: $brand-info;
-			}
-		}
-	}
-}
-
-.page-header {
-	&--order {
-		margin-top: $grid-gutter-width/3;
-	}
-}
-
-.js {
-	.js-hidden {
-		display: none;
-	}
-}
-
-.nav-pills {
-	&.nav-wizard {
-    display: table;
-    a {
-      position: relative;
-			overflow: visible;
-			border-right: 15px solid $body-bg;
-			border-left: 15px solid $body-bg;
-      display: block;
-      width: 100%;
-      padding-top: 10px;
-      margin-top: 5px;
-      @media (min-width: $screen-xs-max) {
-        display: table-cell;
-        width: 25%;
-        padding-top: 0;
-        vertical-align: middle;
-      }
-      border-radius: 0;
-      background-color: $gray-lighter;
-      height: 40px;
-      text-align: center;
-      
-      &:first-of-type {
-        border-left: 0;
-        border-radius:5px 0 0 5px;
-      }
-      
-      &:last-of-type {
-        border-right: 0;
-        border-radius:0 5px 5px 0;
-      }
-      
-      &:hover {
-        text-decoration: none;
-      }
-    }
-
-    .nav-arrow {
-      position: absolute;
-      top: 0px;
-      right: -20px;
-      width: 0px;
-      height: 0px;
-      border-style: solid;
-      border-width: 20px 0 20px 20px;
-      border-color: transparent transparent transparent $gray-lighter;
-      z-index: 150;
-    }
-
-    .nav-wedge {
-      position: absolute;
-      top: 0px;
-      left: -20px;
-      width: 0px;
-      height: 0px;
-      border-style: solid;
-      border-width: 20px 0 20px 20px;
-      border-color: $gray-lighter $gray-lighter $gray-lighter transparent;
-      z-index: 150;
-    }
-
-    .active {
-      .nav-arrow {
-        border-color: transparent transparent transparent $nav-pills-active-link-hover-bg;
-      }
-
-      .nav-wedge {
-        border-color: $nav-pills-active-link-hover-bg $nav-pills-active-link-hover-bg $nav-pills-active-link-hover-bg transparent;
-      }
-
-        background-color: $nav-pills-active-link-hover-bg;
-        color: $nav-pills-active-link-hover-color;
-    }
-	}
-}
