+++ saleor/static/scss/storefront/components/_cart-dropdown.scss
