+++ saleor/static/scss/storefront/components/_cart.scss
