+++ saleor/static/scss/storefront/components/_checkout.scss
