+++ saleor/static/scss/storefront/components/_footer.scss
