+++ saleor/static/scss/storefront/components/_header.scss
