+++ saleor/static/scss/storefront/components/_home.scss
