+++ saleor/static/scss/storefront/components/_labels.scss
