+++ saleor/static/scss/storefront/components/_pagination.scss
