+++ saleor/static/scss/storefront/components/_product-list.scss
