+++ saleor/static/scss/storefront/components/_product-page.scss
