+++ templates/base.html
@@ -14,7 +14,7 @@
         {% endblock meta %}
 
         {% render_bundle 'vendor' 'css' %}
-        {% render_bundle 'demo' 'css' %}
+        {% render_bundle 'storefront' 'css' %}
 
         {% block stylesheet %}{% endblock stylesheet %}
 
@@ -166,7 +166,7 @@
         {% endblock %}
         {% block footer_scripts %}
             {% render_bundle 'vendor' 'js' %}
-            {% render_bundle 'demo' 'js' %}
+            {% render_bundle 'storefront' 'js' %}
         {% endblock footer_scripts %}
     </body>
 </html>
