+++ webpack.config.js
@@ -28,10 +28,8 @@ var providePlugin = new webpack.ProvidePlugin({
 
 var config = {
   entry: {
-    cart: './saleor/static/js/cart.js',
     dashboard: './saleor/static/js/dashboard.js',
     storefront: './saleor/static/js/storefront.js',
-    demo: './saleor/static/js/demo/storefront.js',
     vendor: [
       'bootstrap',
       'jquery',
